Update 02.js
'const' cannot change; never change. Therefore, const a = 4; is creating a constant variable named 'a' and initializes it with the value of '4'. Once a constant variable is set and initialized under 'const', its value cannot be changed. The next line 'a=5' tries to change its value of 'a' to '5', which is not allowed for a constant variable in JS. So, to fix this error, 'a' can be declared as a regular variable using 'let', instead of 'const'.